requirements.txt
torch
torchvision
transformers
opencv-python
pillow
pyttsx3
flask
gunicorn
